separate NCAxis to partial class for statuses
