+++ TwinCat Motion ADS/NcAxisStatuses.cs
@@ -1,0 +1,124 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace TwinCat_Motion_ADS
+{
+    public partial class NcAxis
+    {
+        //Axis error
+        private bool _error;
+        public bool Error
+        {
+            get { return _error; }
+            set { _error = value; OnPropertyChanged(); }
+        }
+
+        //Axis position (encoder)
+        private double _axisPosition;
+        public double AxisPosition
+        {
+            get { return _axisPosition; }
+            set { _axisPosition = value; OnPropertyChanged(); }
+        }
+        //Axis enabled status
+        private bool _axisEnabled;
+        public bool AxisEnabled
+        {
+            get { return _axisEnabled; }
+            set { _axisEnabled = value; OnPropertyChanged(); }
+        }
+        //Axis forward enabled status
+        private bool _axisFwEnabled;
+        public bool AxisFwEnabled
+        {
+            get { return _axisFwEnabled; }
+            set { _axisFwEnabled = value; OnPropertyChanged(); }
+        }
+        //Axis backward enabled status
+        private bool _axisBwEnabled;
+        public bool AxisBwEnabled
+        {
+            get { return _axisBwEnabled; }
+            set { _axisBwEnabled = value; OnPropertyChanged(); }
+        }
+
+        private async Task<bool> checkFwLimitTask(bool limitStatus, CancellationToken wToken, int msDelay = 50)
+        {
+            //if limitStatus true, watch for FwEnabled flag to go false and complete the task
+            while (await read_bFwEnabled() == limitStatus)
+            {
+                if (wToken.IsCancellationRequested)
+                {
+                    throw new TaskCanceledException();
+                }
+                await Task.Delay(msDelay);
+            }
+            return true;
+        }
+
+        private async Task<bool> checkBwLimitTask(bool limitStatus, CancellationToken wToken, int msDelay = 50)
+        {
+            //if limitStatus true, watch for BwEnabled flag to go false and complete the task
+            while (await read_bBwEnabled() == limitStatus)
+            {
+                if (wToken.IsCancellationRequested)
+                {
+                    throw new TaskCanceledException();
+                }
+                await Task.Delay(msDelay);
+            }
+            return true;
+        }
+
+        public async Task<bool> waitForDone(CancellationToken wToken)
+        {
+            bool doneStatus = await read_bDone();
+            while (doneStatus != true)
+            {
+                doneStatus = await read_bDone();
+                //Cancellation method for task otherwise task will never complete if no done signal ever received
+                if (wToken.IsCancellationRequested)
+                {
+                    throw new TaskCanceledException();
+                }
+            }
+            return true;
+        }
+
+        public async Task<bool> checkForError(CancellationToken wToken)
+        {
+            bool errorStatus = await read_bError();
+            while (errorStatus != true)
+            {
+                errorStatus = await read_bError();
+                if (wToken.IsCancellationRequested)
+                {
+                    throw new TaskCanceledException();
+                }
+            }
+            Console.WriteLine("Axis error");
+            return true;
+        }
+
+        public async Task<bool> read_bDone()
+        {
+            var result = await Plc.TcAds.ReadAnyAsync<bool>(bDoneHandle, CancellationToken.None);
+            return result.Value;
+        }
+
+        public async Task<double> read_AxisPosition()
+        {
+            var result = await Plc.TcAds.ReadAnyAsync<double>(fActPositionHandle, CancellationToken.None);
+            AxisPosition = result.Value;
+            return result.Value;
+        }
+
+
+
+
+    }
+}
